feat: Add file extension validation for input raster layer
This commit adds file extension validation for the input raster layer in the `ChloeAlgorithm` class. The `checkParameterValues` method now checks if the selected raster layer from the canvas is a supported file extension (GeoTIFF or ASCII). If the selected raster layer has a different file extension, an error message is displayed to the user.